WIP
Todo:
[] - Use the symfony driver (browserkit) as default driver to run the
tests on. Now we seem to use Selenium2 with Chrome by default. And that
causes some of the tests to fail. Wayf tetss for example are unable to
click certain links because they are not on the page?
[] - Sessions are not stable yet. The authentication state is lost
regularly. Seems to relate to the loop guard section that is commented
out in: src/OpenConext/EngineBlockBundle/EventListener/AuthenticationStateInitializer.php
[] - I started running the different features, and fixed some issues
here and there. But I suggest to use the failing tests to find/fix the
remaining issues present in the EB upgrade